refactor:code splitted into reducers and utilities
