+++ src/reducers/index.js
@@ -1,1 +1,2 @@
 export { noteInputReducer, defaultNoteData } from "./NoteInputReducer";
+export { initialNoteFilterState, noteFilterReducer } from "./NoteFilterReducer";
