+++ src/utilities/index.js
@@ -1,2 +1,3 @@
 export { fetchAndSetFormInputValues } from "./fetchAndSetFormInputValues";
-export * from "./dateFormattingUtilities.js";
+export * from "./dateFormattingUtilities";
+export * from "./noteFilterFunctions";
